Factor common functionality in traits for schema builders
