+++ src/backend/sqlite/index.rs
@@ -40,9 +40,9 @@ impl IndexBuilder for SqliteQueryBuilder {
             table.prepare(sql, '`');
         }
     }
-}
 
-impl SqliteQueryBuilder {
+    fn write_column_index_prefix(&self, _col_prefix: &Option<u32>, _sql: &mut SqlWriter) {}
+
     fn prepare_index_prefix(&self, create: &IndexCreateStatement, sql: &mut SqlWriter) {
         if create.primary {
             write!(sql, "PRIMARY ").unwrap();
@@ -52,27 +52,4 @@ impl SqliteQueryBuilder {
         }
     }
 
-    fn prepare_index_name(&self, name: &Option<String>, sql: &mut SqlWriter) {
-        if let Some(name) = name {
-            write!(sql, "`{}`", name).unwrap();
-        }
-    }
-
-    fn prepare_index_columns(&self, columns: &[IndexColumn], sql: &mut SqlWriter) {
-        write!(sql, " (").unwrap();
-        columns.iter().fold(true, |first, col| {
-            if !first {
-                write!(sql, ", ").unwrap();
-            }
-            col.name.prepare(sql, '`');
-            if let Some(order) = &col.order {
-                match order {
-                    IndexOrder::Asc => write!(sql, " ASC").unwrap(),
-                    IndexOrder::Desc => write!(sql, " DESC").unwrap(),
-                }
-            }
-            false
-        });
-        write!(sql, ")").unwrap();
-    }
 }
