+++ src/backend/sqlite/mod.rs
@@ -18,3 +18,5 @@ impl Default for SqliteQueryBuilder {
 impl GenericBuilder for SqliteQueryBuilder {}
 
 impl SchemaBuilder for SqliteQueryBuilder {}
+
+impl QuotedBuilder for SqliteQueryBuilder {}
